+++ crates/entrypoint/src/cli/command/azure/entra/mod.rs
@@ -1,12 +1,14 @@
 pub mod azure_entra;
 pub mod service_principal;
 pub mod user;
+pub mod group;
 
 pub use azure_entra::AzureEntraCommand;
 use clap::Args;
 use eyre::Result;
 pub use service_principal::AzureEntraServicePrincipalArgs;
 pub use user::AzureEntraUserArgs;
+pub use group::AzureEntraGroupArgs;
 
 /// Entra (Azure AD) related commands.
 #[derive(Args, Debug, Clone)]